+++ ffi/posix_types_x86_h.lua
@@ -1,0 +1,26 @@
+local ffi = require("ffi")
+
+ffi.cdef[[
+typedef unsigned int size_t;
+typedef int ssize_t;
+typedef long int off_t;
+struct timeval {
+  long int tv_sec;
+  long int tv_usec;
+};
+struct statvfs {
+  long unsigned int f_bsize;
+  long unsigned int f_frsize;
+  long unsigned int f_blocks;
+  long unsigned int f_bfree;
+  long unsigned int f_bavail;
+  long unsigned int f_files;
+  long unsigned int f_ffree;
+  long unsigned int f_favail;
+  long unsigned int f_fsid;
+  int __f_unused;
+  long unsigned int f_flag;
+  long unsigned int f_namemax;
+  int __f_spare[6];
+};
+]]
